+++ packages/flutter_app_publisher/lib/src/flutter_app_publisher.dart
@@ -11,6 +11,7 @@ class FlutterAppPublisher {
     AppPackagePublisherFirebase(),
     AppPackagePublisherFirebaseHosting(),
     AppPackagePublisherGithub(),
+    AppPackagePublisherMinio(),
     AppPackagePublisherPgyer(),
     AppPackagePublisherPlayStore(),
     AppPackagePublisherQiniu(),