+++ packages/flutter_app_publisher/lib/src/publishers/publishers.dart
@@ -4,6 +4,7 @@ export 'fir/app_package_publisher_fir.dart';
 export 'firebase/app_package_publisher_firebase.dart';
 export 'firebase_hosting/app_package_publisher_firebase_hosting.dart';
 export 'github/app_package_publisher_github.dart';
+export 'minio/app_package_publisher_minio.dart';
 export 'pgyer/app_package_publisher_pgyer.dart';
 export 'playstore/app_package_publisher_playstore.dart';
 export 'qiniu/app_package_publisher_qiniu.dart';