+++ packages/unified_distributor/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.3
+
+* feat: Support `minio` publisher
+
 ## 0.2.2
 
 * feat: Support `app-version` argument to override version from pubspec.yaml (future versions will not read from pubspec.yaml)